pythondashboard/requirements.txt
streamlit==1.31.0
pandas==2.2.0
requests==2.31.0
plotly==5.18.0
streamlit-plotly-events==0.0.6
python-dotenv==1.0.0
